navbar refactor. Split up Navbar.tsx into separate files and renamed components to make more readable
